fix (behavior): relocation of file
