+++ tests/engine/core/behavior.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include <engine/core/behavior.h>
+#include <engine/public/behavior.h>
 #include <engine/public/scene.h>
 #include <engine/public/component.h>
 #include <engine/public/components/behaviorscript.h>
